refactored the kernel spec classes and created skeleton code to start the test tool
